+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 Version 0.18.0 *(2021-09-13)*
 ---------------------------------
 
+- **Behavior Change:** When using version 7.1.0 or newer of the Android Gradle Plugin we will now publish all variants
+of a library unless `androidVariantToPublish` was set in the DSL. This means that for example both `debug` and `release`
+or all flavors.
+- Deprecated `androidVariantToPublish`. In the future the main plugin will always publish all variants of an Android
+library. If you need to publish only one variant or a subset take a look at the [base plugin](../README.md#base-plugin)
+APIs.
+- Base plugin: Added `AndroidSingleVariantLibrary` and `AndroidMultiVariantLibrary` options that use the new AGP 7.1
+APIs under the hood.
+- Base plugin: Deprecated `AndroidLibrary` option in favor of the above
+
+
+Version 0.18.0 *(2021-09-13)*
+---------------------------------
+
 - The minimum supported Kotlin version is now `1.4.30`
 - It's now possible to specify `SONATYPE_HOST` as a Gradle property, e.g.
   - `SONATYPE_HOST=S01` for `s01.sonatype.org`